ImmigrationAI/scraper/requirements.txt
scrapy==2.11.2
scrapy-playwright==0.0.33
playwright==1.44.0
openai>=1.52.0
httpx==0.27.0
qdrant-client==1.9.1
psycopg2-binary==2.9.9
langchain-text-splitters==0.2.0
apscheduler==3.10.4
python-dotenv==1.0.1
pdfplumber==0.11.0
requests==2.32.3
prometheus-client>=0.20.0